requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
google-adk==0.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
httpx>=0.25.2
pydantic>=2.5.0
python-multipart>=0.0.6
twilio>=8.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
